refactor: 💡 Create separate Offerings Controller
Also addressed review changes and updated swagger docs
