+++ src/assets/assets.controller.spec.ts
@@ -5,9 +5,6 @@ import {
   ConditionType,
   KnownTokenType,
   ScopeType,
-  StoBalanceStatus,
-  StoSaleStatus,
-  StoTimingStatus,
   TokenIdentifierType,
 } from '@polymathnetwork/polymesh-sdk/types';
 
@@ -15,8 +12,6 @@ import { AssetsController } from '~/assets/assets.controller';
 import { AssetsService } from '~/assets/assets.service';
 import { PaginatedResultsModel } from '~/common/models/paginated-results.model';
 import { ResultsModel } from '~/common/models/results.model';
-import { OfferingsService } from '~/offerings/offerings.service';
-import { createOfferingDetailsModel } from '~/offerings/offerings.util';
 import { MockSecurityTokenClass } from '~/test-utils/mocks';
 
 describe('AssetsController', () => {
@@ -30,19 +25,13 @@ describe('AssetsController', () => {
     findTrustedClaimIssuers: jest.fn(),
   };
 
-  const mockOfferingsService = {
-    findAllByTicker: jest.fn(),
-  };
-
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       controllers: [AssetsController],
-      providers: [AssetsService, OfferingsService],
+      providers: [AssetsService],
     })
       .overrideProvider(AssetsService)
       .useValue(mockAssetsService)
-      .overrideProvider(OfferingsService)
-      .useValue(mockOfferingsService)
       .compile();
 
     controller = module.get<AssetsController>(AssetsController);
@@ -211,66 +200,6 @@ describe('AssetsController', () => {
       const result = await controller.getTrustedClaimIssuers({ ticker: 'SOME_TICKER' });
 
       expect(result).toEqual({ results: mockClaimIssuers });
-    });
-  });
-
-  describe('getOfferings', () => {
-    it('should return the list of Offerings for an Asset', async () => {
-      const mockOfferings = [
-        {
-          sto: {
-            id: new BigNumber('1'),
-          },
-          details: {
-            tiers: [
-              {
-                amount: new BigNumber('1000'),
-                price: new BigNumber('1'),
-                remaining: new BigNumber('1000'),
-              },
-            ],
-            creator: {
-              did: 'Ox6'.padEnd(66, '0'),
-            },
-            name: 'SERIES A',
-            offeringPortfolio: {
-              did: 'Ox6'.padEnd(66, '0'),
-            },
-            raisingPortfolio: {
-              did: 'Ox6'.padEnd(66, '0'),
-            },
-            raisingCurrency: 'CURRENCY',
-            venue: {
-              id: new BigNumber('1'),
-            },
-            start: new Date(),
-            end: null,
-            status: {
-              timing: StoTimingStatus.Started,
-              balance: StoBalanceStatus.Available,
-              sale: StoSaleStatus.Live,
-            },
-            minInvestment: new BigNumber('1'),
-            totalAmount: new BigNumber('1000'),
-            totalRemaining: new BigNumber('1000'),
-          },
-        },
-      ];
-
-      mockOfferingsService.findAllByTicker.mockResolvedValue(mockOfferings);
-
-      const result = await controller.getOfferings(
-        { ticker: 'SOME_TICKER' },
-        { timing: StoTimingStatus.Started }
-      );
-
-      const mockResult = new ResultsModel({
-        results: mockOfferings.map(offering =>
-          // eslint-disable-next-line @typescript-eslint/no-explicit-any
-          createOfferingDetailsModel(offering as any)
-        ),
-      });
-      expect(result).toEqual(mockResult);
     });
   });
 });
