+++ src/assets/dto/ticker-params.dto.ts
@@ -1,0 +1,8 @@
+/* istanbul ignore file */
+
+import { IsTicker } from '~/common/decorators/validation';
+
+export class TickerParamsDto {
+  @IsTicker()
+  readonly ticker: string;
+}
